move error models; fix tslint errors
