+++ __tests__/errors.test.ts
@@ -1,4 +1,4 @@
-import { Repository } from "typeorm"
+import { Repository } from 'typeorm'
 import { ErrorLogic } from '../src/logic/errors';
 import { JSError } from '../src/data/entities/error';
