+++ src/models/errors/bad-request.ts
@@ -1,4 +1,4 @@
-import { BaseError } from "./base-error";
+import { BaseError } from './base-error';
 
 export class BadRequestError extends BaseError {
     constructor(message: string, status = 400, err?: Error) {
