+++ src/models/errors/base-error.ts
