+++ src/models/errors/forbidden.ts
@@ -1,4 +1,4 @@
-import { BaseError } from "./base-error";
+import { BaseError } from './base-error';
 
 export class ForbiddenError extends BaseError {
     constructor(message: string, status = 403, err?: Error) {
