+++ src/models/errors/not-found.ts
@@ -1,4 +1,4 @@
-import { BaseError } from "./base-error";
+import { BaseError } from './base-error';
 
 export class NotFoundError extends BaseError {
     constructor(message: string, status = 404, err?: Error) {
