+++ src/models/errors/unauthorized.ts
@@ -1,4 +1,4 @@
-import { BaseError } from "./base-error";
+import { BaseError } from './base-error';
 
 export class UnauthorizedError extends BaseError {
     constructor(message: string, status = 401, err?: Error) {
